requirements.txt
Jinja2==3.0.3
rdflib>=6.2.0
WTForms
starlette
starlette-wtf
jinja2
fastapi
uvicorn
pyshacl
pyyaml
requests
pydantic>=2.0.0
pydantic_settings

# Test dependencies
pytest>=7.0.0
pytest-logging>=2015.11.4
